refactor: drop 0.3 compat
