+++ gateway/Cargo.toml
@@ -4,10 +4,6 @@ version = "0.1.0"
 edition = "2021"
 publish = false
 
-[features]
-compat_v0_3 = []
-default = [ "compat_v0_3" ]
-
 [dependencies]
 serde = { version = "1.0.137", features = [ "derive" ] }
 serde_json = "1.0.81"
