+++ gateway/src/api/latest.rs
