+++ gateway/src/api/mod.rs
@@ -1,5 +1,2 @@
-#[cfg(feature = "compat_v0_3")]
-pub mod v0_3;
-
-pub mod v0_4;
-pub use v0_4::make_api;
+pub mod latest;
+pub use latest::make_api;
